Create FundAgeBalances
This will look at a participant in the 401k only, create a tag based their age rage, and identify if they are in just one fund or multiple funds (Source does not matter).

Then if they are in a Target date fund, make a tag and determine if they are in one within 7 years of expected retirement age, call it correct. Otherwise it is incorrect.